+++ packages/backend-ai/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @aws-amplify/backend-ai
 
+## 0.3.4
+
+### Patch Changes
+
+- Updated dependencies [63fb254]
+  - @aws-amplify/ai-constructs@0.7.0
+
+## 0.3.3
+
+### Patch Changes
+
+- bd4ff4d: Add memory setting to conversation handler
+- 0d6489d: Use AiModel from data-schema-types as possible input
+- Updated dependencies [5f46d8d]
+- Updated dependencies [bd4ff4d]
+  - @aws-amplify/backend-output-schemas@1.4.0
+  - @aws-amplify/ai-constructs@0.6.2
+
 ## 0.3.2
 
 ### Patch Changes